Refactor code out of main module
